% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{apply_dco}
\alias{apply_dco}
\title{Apply data cut-off}
\usage{
apply_dco(df, dco = NULL, events = NULL)
}
\arguments{
\item{df}{Data frame with uncensored data. Generated from \code{sim_t_uncensored}}

\item{dco}{Data cut-off time. In time-units since start of the trial.}

\item{events}{May be specified instead of \code{dco}. The number of events that triggers data cut-off.}
}
\value{
A data frame containing simulated data. Snapshot taken at data cut-off.
}
\description{
Apply a data cut-off to an uncensored data set.
}
