% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{crit_1_of_2}
\alias{crit_1_of_2}
\title{Critical value at interim analysis}
\usage{
crit_1_of_2(var_u_int, design, alpha_spend_f = ldobf,
  alpha_one_sided = 0.025)
}
\arguments{
\item{var_u_int}{The observed variance of the weighted log-rank statistic at the interim.}

\item{design}{The design object, created using \code{two_stage_design}}

\item{alpha_spend_f}{The alpha-spending function. Default is the Lan-DeMets O'Brien-Fleming function.}

\item{alpha_one_sided}{One-sided alpha level.}
}
\value{
The critical value for the interim z-statistic.
}
\description{
Find the critical value at the interim analysis of a two-stage design
}
