% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{crit_1_of_3}
\alias{crit_1_of_3}
\title{Critical value at first interim analysis}
\usage{
crit_1_of_3(var_u_int_1, design, alpha_spend_f = ldobf,
  alpha_one_sided = 0.025)
}
\arguments{
\item{var_u_int_1}{The observed variance of the weighted log-rank statistic at the first interim.}

\item{design}{The design object, created using \code{three_stage_design}}

\item{alpha_spend_f}{The alpha-spending function. Default is the Lan-DeMets O'Brien-Fleming function.}

\item{alpha_one_sided}{One-sided alpha level.}
}
\value{
The critical value for the first interim z-statistic.
}
\description{
Find the critical value at the first interim analysis of a three-stage design
}
