% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{crit_2_of_3}
\alias{crit_2_of_3}
\title{Critical value at second interim analysis}
\usage{
crit_2_of_3(info_frac_sf, info_frac_1_2, crit_1, design,
  alpha_spend_f = ldobf, alpha_one_sided = 0.025)
}
\arguments{
\item{info_frac_sf}{Information fraction to be used in the spending function at the second interim.}

\item{info_frac_1_2}{The observed information fraction: determines covariance of test statistics.}

\item{crit_1}{The critical value for the z-statistic at the first interim, found via \code{crit_1_of_3}.}

\item{alpha_spend_f}{The alpha-spending function. Default is the Lan-DeMets O'Brien-Fleming function.}

\item{alpha_one_sided}{One-sided alpha level.}

\item{var_u_int_1}{The observed variance of the weighted log-rank statistic at the first interim.}
}
\value{
The critical value for the second interim z-statistic.
}
\description{
Find the critical value at the second interim analysis of a three-stage design
}
