% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_events.R
\name{expected_dco_two_arm}
\alias{expected_dco_two_arm}
\title{Expected time of dco}
\usage{
expected_dco_two_arm(total_events, recruitment, model)
}
\arguments{
\item{{total_events}}{Number of events.}

\item{{recruitment}}{List of recruitment information.
Containing \enumerate{
              \item Sample size on control, \code{n_0}
              \item Sample size on experimental, \code{n_1}
              \item Recruitment period, \code{r_period}
              \item Recruitment parameter for power model, \code{k}
            }}

\item{{model}}{The piecewise hazard model.
A list containing the \code{change_points} and \code{lambdas}.}
}
\value{
The expected calendar time of reaching \code{total_events}.
}
\description{
\code{expected_dco_two_arm} returns the expected calendar time of data-cut-off for a given number of events.
}
