% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_events.R
\name{expected_events_two_arm}
\alias{expected_events_two_arm}
\title{Expected number of events}
\usage{
expected_events_two_arm(dco = 28, recruitment, model,
  total_only = TRUE)
}
\arguments{
\item{{dco}}{Time of data cut-off.}

\item{{recruitment}}{List of recruitment information.
Containing \enumerate{
              \item Sample size on control, \code{n_0}
              \item Sample size on experimental, \code{n_1}
              \item Recruitment period, \code{r_period}
              \item Recruitment parameter for power model, \code{k}
            }}

\item{{model}}{The piecewise hazard model.
A list containing the \code{change_points} and \code{lambdas}.}

\item{{total_only}}{By default, the function returns only the expected number of events.
To see more detailed output set \code{total_only = FALSE}.}
}
\value{
The expected number of events at \code{dco}.
}
\description{
\code{expected_events_two_arm} returns the expected number of events at a given data-cut-off time.
}
