% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futility.R
\name{futility_three_stage}
\alias{futility_three_stage}
\title{Futility three-stage design}
\usage{
futility_three_stage(design, hr_threshold_1 = 1.2,
  hr_threshold_2 = 1.1)
}
\arguments{
\item{design}{The design object, created using \code{three_stage_design}}

\item{hr_threshold_1}{The threshold for futility at the first interim, on the (average) HR scale}

\item{hr_threshold_d}{The threshold for futility at the second interim, on the (average) HR scale}
}
\value{
The design object, with futility boundary and updated power/expected duration.
}
\description{
Create a three-stage design by adding a futility boundary to an existing three-stage design
}
