% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futility.R
\name{futility_two_stage}
\alias{futility_two_stage}
\title{Futility two-stage design}
\usage{
futility_two_stage(design, hr_threshold)
}
\arguments{
\item{design}{The design object, created using \code{two_stage_design}}

\item{hr_threshold}{The threshold for futility on the (average) HR scale}
}
\value{
The design object, with futility boundary and updated power/expected duration.
}
\description{
Create a two-stage design by adding a futility boundary to an existing two-stage design
}
