% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_events.R
\name{get_t_pieces}
\alias{get_t_pieces}
\title{Time-at-risk intervals.}
\usage{
get_t_pieces(t, ts)
}
\arguments{
\item{{t}}{Total time.}

\item{{ts}}{The breakpoints for intervals, starting at 0, ending at Inf.}
}
\value{
A vector of length one less than \code{ts}, containing time at risk between consecutive breakpoints in \code{ts}.
}
\description{
\code{get_t_pieces} returns the time at risk between intervals \code{ts} up to time \code{t}.
}
