% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{s_to_t_star}
\alias{s_to_t_star}
\title{Convert from s_star to t_star based on model assumptions}
\usage{
s_to_t_star(s_star, recruitment, model)
}
\arguments{
\item{s_star}{The s_star parameter of the modestly weighted logrank test. A number between 0 and 1.}

\item{model}{The piecewise hazard model.
A list containing the \code{change_points} and \code{lambdas}.}
}
\value{
The t_star parameter of the modestly weighted logrank test
}
\description{
Find the value of t_star that corresponds to a given s_star, given model assumptions.
}
