% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{sim_t_uncensored}
\alias{sim_t_uncensored}
\title{Simulate uncensored data}
\usage{
sim_t_uncensored(model, recruitment)
}
\arguments{
\item{model}{The piecewise hazard model.
A list containing the \code{change_points} and \code{lambdas}.}

\item{recruitment}{List of recruitment information.
Containing \enumerate{
              \item Sample size on control, \code{n_0}
              \item Sample size on experimental, \code{n_1}
              \item Recruitment period, \code{r_period}
              \item Recruitment parameter for power model, \code{k}
            }}

\item{var_u_int}{The observed variance of the weighted log-rank statistic at the interim.}
}
\value{
A data frame containing simulated data.
}
\description{
Simulate data as if there is no administrative censoring, i.e., all events observed.
}
