% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{stagewise_p_2}
\alias{stagewise_p_2}
\title{Stage-wise p-value at 2nd analysis}
\usage{
stagewise_p_2(crit_1, z_2, v_1, v_2)
}
\arguments{
\item{crit_1}{Critical value at first analysis. Expecting a negative number.}

\item{z_2}{The observed z-statistic at the second analysis..}

\item{v_1}{The observed variance of the u-statistic at the first analysis.}

\item{v_2}{The observed variance of the u-statistic at the second analysis.}
}
\description{
Find the stage-wise one-sided value, when a trial has crossed efficacy boundary at the second analysis.
}
