% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{stagewise_p_3}
\alias{stagewise_p_3}
\title{Stage-wise p-value at 3rd analysis}
\usage{
stagewise_p_3(crit_1, crit_2, z_3, v_1, v_2, v_3)
}
\arguments{
\item{crit_1}{Critical value at first analysis. Expecting a negative number.}

\item{crit_2}{Critical value at second analysis. Expecting a negative number.}

\item{z_3}{The observed z-statistic at the third analysis..}

\item{v_1}{The observed variance of the u-statistic at the first analysis.}

\item{v_2}{The observed variance of the u-statistic at the second analysis.}

\item{v_3}{The observed variance of the u-statistic at the third analysis.}
}
\description{
Find the stage-wise one-sided value, when a trial has crossed efficacy boundary at the third analysis.
}
