% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_events.R
\name{surv_pieces_simple}
\alias{surv_pieces_simple}
\title{Probability of event-free at time t, under piecewise constant hazard.}
\usage{
surv_pieces_simple(t, change_points, lambdas)
}
\arguments{
\item{{t}}{time}

\item{{change_points}}{The breakpoints for the piecewise hazard function.}

\item{{lambdas}}{A vector of length one more than \code{change_points}. The hazard between consecutive breakpoints in \code{c(0, change_points, Inf)}.}
}
\value{
The probability of being free of the event-of-interest at time \code{t}.
}
\description{
\code{surv_pieces} returns the probability of being event-free at time \code{t} assuming piecewise constant hazard.
}
