% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three_stage_design.R
\name{three_stage_design}
\alias{three_stage_design}
\title{Three-stage design}
\usage{
three_stage_design(t_star, model, recruitment, dco_int_1, dco_int_2,
  dco_final, events_int_1 = NULL, events_int_2 = NULL,
  events_final = NULL, alpha_one_sided = 0.025,
  alpha_spend_f = ldobf, length_t)
}
\arguments{
\item{t_star}{Parameter of the modestly-weighted log-rank test. Setting t_star=0 corresponds to a standard log-rank test.}

\item{model}{The piecewise hazard model.
A list containing the \code{change_points} and \code{lambdas}.}

\item{recruitment}{List of recruitment information.
Containing \enumerate{
              \item Sample size on control, \code{n_0}
              \item Sample size on experimental, \code{n_1}
              \item Recruitment period, \code{r_period}
              \item Recruitment parameter for power model, \code{k}
            }}

\item{dco_int_1}{The time of the first interim analysis in time-units since start of the trial.}

\item{dco_int_2}{The time of the second interim analysis in time-units since start of the trial.}

\item{dco_final}{The time of the final analysis in time-units since start of the trial.}

\item{events_int_1}{May be specified instead of \code{dco_int_1}. The number of events that triggers the first interim analysis.}

\item{events_int_2}{May be specified instead of \code{dco_int_2}. The number of events that triggers the second interim analysis.}

\item{events_final}{May be specified instead of \code{dco_final}. The number of events that triggers the final analysis.}

\item{alpha_one_sided}{One-sided alpha level.}

\item{length_t}{Number of cutpoints to use when approximating the distribution of the MWLRT-statistic. Default is 18. Can be increased for greater accuracy.}

\item{alpha_spending_f}{The alpha-spending function. Default is the Lan-DeMets O'Brien-Fleming function.}
}
\value{
A list describing the design.
}
\description{
Three-stage design for a modestly-weighted log-rank test
}
