% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three_stage_sim.R
\name{three_stage_sim}
\alias{three_stage_sim}
\title{Three-stage simulation}
\usage{
three_stage_sim(n_sims = 1, design)
}
\arguments{
\item{n_sims}{The number of simulations.}

\item{design}{The design object, created using \code{three_stage_design}}
}
\value{
A data-frame containing: critical values, z-statistics, timing of analyses.
}
\description{
Simulate a three-stage trial, analyzed with a modestly-weighted log-rank test
}
