% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_stage_design.R
\name{two_stage_design}
\alias{two_stage_design}
\title{Two-stage design}
\usage{
two_stage_design(t_star, model, recruitment, dco_int, dco_final,
  events_int = NULL, events_final = NULL, alpha_one_sided = 0.025,
  alpha_spend_f = ldobf, length_t = 18)
}
\arguments{
\item{t_star}{Parameter of the modestly-weighted log-rank test. Setting t_star=0 corresponds to a standard log-rank test.}

\item{model}{The piecewise hazard model.
A list containing the \code{change_points} and \code{lambdas}.}

\item{recruitment}{List of recruitment information.
Containing \enumerate{
              \item Sample size on control, \code{n_0}
              \item Sample size on experimental, \code{n_1}
              \item Recruitment period, \code{r_period}
              \item Recruitment parameter for power model, \code{k}
            }}

\item{dco_int}{The time of the interim analysis in time-units since start of the trial.}

\item{dco_final}{The time of the final analysis in time-units since start of the trial.}

\item{events_int}{May be specified instead of \code{dco_int}. The number of events that triggers the interim analysis.}

\item{events_final}{May be specified instead of \code{dco_final}. The number of events that triggers the final analysis.}

\item{alpha_one_sided}{One-sided alpha level.}

\item{length_t}{Number of cutpoints to use when approximating the distribution of the MWLRT-statistic. Default is 18. Can be increased for greater accuracy.}

\item{alpha_spending_f}{The alpha-spending function. Default is the Lan-DeMets O'Brien-Fleming function.}
}
\value{
A list describing the design.
}
\description{
Two-stage design for a modestly-weighted log-rank test
}
