% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{crit_1_of_2}
\alias{crit_1_of_2}
\title{Critical value at interim analysis}
\usage{
crit_1_of_2(info_frac_sf, alpha_spend_f = ldobf,
  alpha_one_sided = 0.025)
}
\arguments{
\item{info_frac_sf}{Information fraction to be used in the spending function at the first interim.}

\item{alpha_spend_f}{The alpha-spending function. Default is the Lan-DeMets O'Brien-Fleming function.}

\item{alpha_one_sided}{One-sided alpha level.}
}
\value{
The critical value for the interim z-statistic.
}
\description{
Find the critical value at the interim analysis of a two-stage design
}
