% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{crit_2_of_2}
\alias{crit_2_of_2}
\title{Critical value at final analysis}
\usage{
crit_2_of_2(info_frac_1_2, crit_1, alpha_spend_f = ldobf,
  alpha_one_sided = 0.025)
}
\arguments{
\item{info_frac_1_2}{The observed information fraction: determines covariance of test statistics.}

\item{crit_1}{The critical value for the z-statistic at the first interim, found via \code{crit_1_of_2}.}

\item{alpha_spend_f}{The alpha-spending function. Default is the Lan-DeMets O'Brien-Fleming function.}

\item{alpha_one_sided}{One-sided alpha level.}
}
\value{
The critical value for the final z-statistic.
}
\description{
Find the critical value at the final analysis of a two-stage design
}
