% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_stage_sim.R
\name{two_stage_sim}
\alias{two_stage_sim}
\title{Two-stage simulation}
\usage{
two_stage_sim(n_sims = 1, design)
}
\arguments{
\item{n_sims}{The number of simulations.}

\item{design}{The design object, created using \code{two_stage_design}}
}
\value{
A data-frame containing: critical values, z-statistics, timing of analyses.
}
\description{
Simulate a two-stage trial, analyzed with a modestly-weighted log-rank test
}
